requirements.txt
Flask==1.0.2
pandas==0.25.1
numpy==1.17.2
google-cloud-storage==1.20.0
scikit-surprise==1.1.0